Consolidate parse, wikidata, precision_display tests together
Every file in tests\*.rs builds a separate binary.  For many of these,
this becomes >200M of executable and symbols, even for a handful of tests.
For build efficiency, we should package the tests together.

I've done this for three of the test binaries, parse, wikidata,
and precision_display, introducing a new miscellaneous_tests package.
I wish to do the same for high_level_functions_tests, list_tests
and svg_rendering_tests, but am leaving those as is for now
until this approach is approved.
