+++ tests/miscellaneous_tests.rs
@@ -1,0 +1,6 @@
+mod miscellaneous_tests
+{
+  mod precision_display;
+  mod wikidata;
+  mod parser;
+}
